# services/api-gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Eucastan_eucastanpay_services_api-gateway_internal_dto_request_user.CreateKYCRequest:
    properties:
      id_number:
        minLength: 11
        type: string
      id_type:
        enum:
        - NIN
        - passport
        - license
        type: string
    required:
    - id_number
    - id_type
    type: object
  github_com_Eucastan_eucastanpay_services_api-gateway_internal_dto_request_user.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  github_com_Eucastan_eucastanpay_services_api-gateway_internal_dto_request_user.RegisterRequest:
    properties:
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password_hash:
        minLength: 8
        type: string
      phone:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password_hash
    type: object
  github_com_Eucastan_eucastanpay_services_api-gateway_internal_dto_request_user.UpdateRequest:
    properties:
      email:
        type: string
      email_verified:
        type: boolean
      first_name:
        type: string
      last_name:
        type: string
      password_hash:
        minLength: 8
        type: string
      phone:
        type: string
      status:
        type: string
    type: object
  github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact:
    name: Stanley Emeh
  description: API Gateway for the EucastanPay Microservices Platform.
  title: EucastanPay API Gateway
  version: "1.0"
paths:
  /{user_id}/kyc:
    patch:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
      security:
      - BearerAuth: []
      summary: Approve user KYC by admin
      tags:
      - KYC
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns JWT tokens.
      parameters:
      - description: Login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_dto_request_user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
      summary: Login user
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new EucastanPay user account.
      parameters:
      - description: User Registration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_dto_request_user.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
      summary: Register a new user
      tags:
      - Authentication
...
  /user:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
      security:
      - BearerAuth: []
      summary: Get User
      tags:
      - User
  /users:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
      security:
      - BearerAuth: []
      summary: List All Users
      tags:
      - User
  /users/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
      security:
      - BearerAuth: []
      summary: Deleting User information
      tags:
      - User
    put:
      consumes:
      - application/json
      parameters:
      - description: Update Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_dto_request_user.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
      security:
      - BearerAuth: []
      summary: Editing User information
      tags:
      - User
  /users/{user_id}/kyc:
    get:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
      security:
      - BearerAuth: []
      summary: Get KYC
      tags:
      - KYC
    post:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: KYC Information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_dto_request_user.CreateKYCRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
      security:
      - BearerAuth: []
      summary: Submit KYC information
      tags:
      - KYC
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"